+++ src/Twilio/Rest/Api/V2010/Account/MessageList.php
@@ -98,6 +98,8 @@ class MessageList extends ListResource
                 Serialize::booleanToString($options['sendAsMms']),
             'ContentVariables' =>
                 $options['contentVariables'],
+            'Tags' =>
+                $options['tags'],
             'RiskCheck' =>
                 $options['riskCheck'],
             'From' =>