+++ src/Twilio/Rest/FlexApi/V1/ConfigurationInstance.php
@@ -71,6 +71,7 @@ use Twilio\Deserialize;
  * @property array[]|null $channelConfigs
  * @property array|null $debuggerIntegration
  * @property array|null $flexUiStatusReport
+ * @property array|null $agentConvEndMethods
  */
 class ConfigurationInstance extends InstanceResource
 {
@@ -131,6 +132,7 @@ class ConfigurationInstance extends InstanceResource
             'channelConfigs' => Values::array_get($payload, 'channel_configs'),
             'debuggerIntegration' => Values::array_get($payload, 'debugger_integration'),
             'flexUiStatusReport' => Values::array_get($payload, 'flex_ui_status_report'),
+            'agentConvEndMethods' => Values::array_get($payload, 'agent_conv_end_methods'),
         ];
 
         $this->solution = [];