+++ CHANGES.md
@@ -1,6 +1,25 @@
 twilio-php Changelog
 ====================
 
+[2023-09-07] Version 7.10.0
+---------------------------
+**Api**
+- Make message tagging parameters public **(breaking change)**
+
+**Flex**
+- Adding `agent_conv_end_methods` to Flex Configuration
+
+**Messaging**
+- Mark Mesasging Services fallback_to_long_code feature obsolete
+
+**Numbers**
+- Add Create Port In request api
+- Renaming sid for bulk_hosting_sid and remove account_sid response field in numbers/v2/BulkHostedNumberOrders **(breaking change)**
+
+**Pricing**
+- gate resources behind a beta_feature
+
+
 [2023-08-24] Version 7.9.0
 --------------------------
 **Api**